Convert extensione2e/kafka tests to REKT (#2065)
* Convert extensione2e/kafka tests to REKT

Except for user permissions tests and kafka source with auth mechanism

* Use testpkg.BootstrapServersPlaintextArr

* Define DELETE_CRD_ON_TEARDOWN flag

Skip deleting CRDs after running tests. Only delete CRDs when manually
calling "make teardown".

Deleting CRDs fails when there are some namespaces still being
terminated after running tests or when some tests keep the namespace
after failure. Deleting CRDs is not really a function of Serverless so
no need to test it in every run.

* Define defaultEnvironment function

It will make sure Knative Service forwarder is used in all scenarios.

* Use the new tag style for knative-eventing-test-eventshub

* Fix deps

+++ hack/lib/serverless.bash
@@ -443,8 +443,10 @@ function teardown_serverless {
   logger.info 'Deleting operators namespace'
   oc delete namespace "${OPERATORS_NAMESPACE}" --ignore-not-found=true
   logger.info 'Ensure not CRDs left'
-  if [[ ! $(oc get crd -oname | grep -c 'knative.dev') -eq 0 ]]; then
-    oc get crd -oname | grep 'knative.dev' | xargs oc delete --timeout=60s
+  if [[ "${DELETE_CRD_ON_TEARDOWN}" == "true" ]]; then
+    if [[ ! $(oc get crd -oname | grep -c 'knative.dev') -eq 0 ]]; then
+      oc get crd -oname | grep 'knative.dev' | xargs oc delete --timeout=60s
+    fi
   fi
   logger.success 'Serverless has been uninstalled.'
 }
